Lab 2_grade_retrieval
Online Grade Retrieval Network Application.
The software to be developed consists of separate server and client applications written using
Python 3. The working software must be demonstrated by each group in one of the laboratory signup
sessions. In the Marking Scheme section below, there is a detailed description of output that the
software must generate when doing the demonstration. The required functionality is described as
follows.